Funcionante no Postman

+++ DDL.sql
@@ -48,7 +48,7 @@ create table tra_trabalho (
   tra_data_hora_entrega datetime not null,
   tra_descricao varchar(200),
   tra_grupo varchar(20) not null,
-  tra_nota int,
+  tra_nota bigint,
   tra_justificativa varchar(100)
 );
